requirements.txt
google-api-python-client==2.31.0
pandas==1.3.4
pyyaml==6.0
matplotlib==3.5.0
numpy
pyti
